MyResftfullApp/Services/UsuarioService/UsuarioService.cs
using MyResftfullApp.Models;
using System.Collections.Generic;
using MyResftfullApp.Repositories;

namespace MyResftfullApp.Services.UsuarioService
{
    public class UsuarioService : IUsuarioService
    {
        private readonly IUsuarioRepository _usuarioRepository;

        public UsuarioService()
        {
            _usuarioRepository = new UsuarioRepository();
        }

        public List<Usuario> ObtenerUsuarios()
        {
            return _usuarioRepository.ObtenerUsuarios();
        }

        public void AgregarUsuario(Usuario usuario)
        {
            _usuarioRepository.AgregarUsuario(usuario);
        }

        public void ModificarUsuario(Usuario usuario)
        {
            _usuarioRepository.ModificarUsuario(usuario);
        }

        public void EliminarUsuario(Usuario usuario)
        {
            _usuarioRepository.EliminarUsuario(usuario);
        }
    }
}